Moved files into src directory for packaging
